refactor: extract CVResult from rolling.py; drop dead get_perf (roadmap 5.2)
Light, sensible split per the audit: only detach what genuinely helps.
- CVResult dataclass -> models/cv_result.py (re-exported from
  fynance.models.rolling and fynance.models; all import paths preserved).
- Remove the dead get_perf() helper (only get_perf2 is used).
Kept intact (cohesive / coupled, splitting would not help): attention.py
(2 coupled classes), econometric_models.py (cohesive functions), and the
_RollingBasis/RollMultiLayerPerceptron inheritance hierarchy.
